+++ tests/tests.py
@@ -380,8 +380,8 @@ class TestFeatureDataset:
     def test_delitem(self, ob_gdb):
         gdb, gdb_path = ob_gdb
         for fds in gdb.values():
-            fcs = fds.feature_classes()
-            for fc_name, fc in fcs:
+            fcs = list(fds.feature_classes().keys())
+            for fc_name in fcs:
                 del fds[fc_name]
             assert len(fds) == 0
 
@@ -444,7 +444,7 @@ class TestFeatureDataset:
     def test_feature_classes(self, ob_gdb):
         gdb, gdb_path = ob_gdb
         for fds in gdb.values():
-            for fc_name, fc in fds.feature_classes():
+            for fc_name, fc in fds.feature_classes().items():
                 assert isinstance(fc_name, str)
                 assert isinstance(fc, ob.FeatureClass)
 
@@ -473,13 +473,13 @@ class TestGeoDatabase:
     def test_delitem(self, ob_gdb):
         gdb, gdb_path = ob_gdb
 
-        for fds_name, fds in gdb.feature_datasets():
-            for fc_name, fc in gdb.feature_classes():
+        for fds_name in list(gdb.feature_datasets().keys()):
+            for fc_name in list(gdb.feature_classes().keys()):
                 try:
                     del gdb[fds_name][fc_name]
                 except KeyError:
                     pass
-            assert len(fds) == 0
+            assert len(gdb[fds_name]) == 0
             del gdb[fds_name]
         assert len(gdb.feature_datasets()) == 0
 
@@ -487,8 +487,8 @@ class TestGeoDatabase:
 
     def test_getitem(self, ob_gdb):
         gdb, gdb_path = ob_gdb
-        for fds_name, fds in gdb.feature_datasets():
-            for fc_name, fc in fds.feature_classes():
+        for fds_name, fds in gdb.feature_datasets().items():
+            for fc_name, fc in fds.feature_classes().items():
                 assert isinstance(gdb[fds_name][fc_name], ob.FeatureClass)
 
         with pytest.raises(KeyError):
@@ -523,7 +523,7 @@ class TestGeoDatabase:
     def test_setitem(self, ob_gdb):
         gdb, gdb_path = ob_gdb
         new_gdb = ob.GeoDatabase()
-        for fds_name, fds in gdb.feature_datasets():
+        for fds_name, fds in gdb.feature_datasets().items():
             new_gdb[fds_name] = fds
             with pytest.raises(KeyError):
                 new_gdb[fds_name] = fds
@@ -534,13 +534,13 @@ class TestGeoDatabase:
 
     def test_feature_classes(self, ob_gdb):
         gdb, gdb_path = ob_gdb
-        for fc_name, fc in gdb.feature_classes():
+        for fc_name, fc in gdb.feature_classes().items():
             assert isinstance(fc_name, str)
             assert isinstance(fc, ob.FeatureClass)
 
     def test_feature_datasets(self, ob_gdb):
         gdb, gdb_path = ob_gdb
-        for fds_name, fds in gdb.feature_datasets():
+        for fds_name, fds in gdb.feature_datasets().items():
             assert isinstance(fds_name, str) or fds_name is None
             assert isinstance(fds, ob.FeatureDataset)
 
@@ -601,7 +601,7 @@ class TestUtilityFunctions:
 
         # noinspection PyUnresolvedReferences
         with pytest.raises(pyogrio.errors.GeometryError):
-            for fc_name, fc in gdb.feature_classes():
+            for fc_name, fc in gdb.feature_classes().items():
                 ob.gdf_to_fc(
                     gdf=fc.to_geodataframe(),
                     gdb_path=gdb_path,
@@ -757,12 +757,12 @@ class TestUsage:
                 assert isinstance(fc_name, str)
                 assert isinstance(fc, ob.FeatureClass)
 
-        for fds_name, fds in gdb.feature_datasets():
+        for fds_name, fds in gdb.feature_datasets().items():
             for fc_name, fc in fds.items():
                 assert isinstance(fc_name, str)
                 assert isinstance(fc, ob.FeatureClass)
 
-        for fc_name, fc in gdb.feature_classes():
+        for fc_name, fc in gdb.feature_classes().items():
             assert isinstance(fc_name, str)
             assert isinstance(fc, ob.FeatureClass)
 
@@ -771,6 +771,6 @@ class TestUsage:
             # noinspection PyTypeChecker
             this_fds = gdb[fds]
             break
-        for fc_name, fc in this_fds.feature_classes():
+        for fc_name, fc in this_fds.feature_classes().items():
             assert isinstance(fc_name, str)
             assert isinstance(fc, ob.FeatureClass)